+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### New Features & Screens
 - Add a Google Cloud Functions that fixes `ITEM_LOGIN_REQUIRED` error from Plaid Link
 - Add a `LinkedAccounts` screen where user can see all the linked accounts grouped by `institution_id`
+- Create a Google Cloud Function that deletes accounts, transactions, and other data related to institution when the user unlinks
   
 ### Bug Fix & Refactor
 - Add a `DateTime` extension that creates a timeago string